requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==4.2.13
django-admin==2.0.2
django-cors-headers==4.4.0
django-excel-response2==3.0.6
django-ratelimit==4.1.0
django-six==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
excel-base==1.0.4
isoweek==1.3.3
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
qrcode==7.4.2
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
xlwt==1.3.0
